Docking API Initial Upload
I have an editor window being added to the editor dock, able to be detached into its own `JFrame`, and get reattached back into the editor dock, but the tab is blank and its title is in the content area of the tab. I am so close to making this work!

-SC